Only show friends on current server.
git-svn-id: http://ecmanaut.googlecode.com/svn/trunk/sites/ikariam.org@195 c5938665-bd24-0410-9958-d597d88377d4

+++ ikariam_friend_list.user.js
@@ -199,7 +199,8 @@ else{
 		
 		flFriendName = IkariamFriend[0];
 		flFriendLink = IkariamFriend[1];
-		
+		if (flFriendLink.indexOf("http://"+ location.host)) continue;
+
 		flHTML += '<li><a href="' + flFriendLink + '">' + flFriendName + '</a> <a href="javascript:flDeleteFriend(\'' + flFriendName + '\', \'' + flFriendLink + '\');"><img src="http://img153.imageshack.us/img153/9549/iconquickdeletech1.gif"></a></li>';
 		
 	}
